[Fix] #45 - jwt 검증 메서드 위치 변경
