+++ src/main/java/com/sascom/chickenstock/global/jwt/JwtProvider.java
@@ -62,24 +62,4 @@ public class JwtProvider {
                 .signWith(key, Jwts.SIG.HS512)
                 .compact();
     }
-
-    public boolean isValidToken(String token) {
-
-        try {
-            Jwts.parser()
-                    .verifyWith(key)
-                    .build()
-                    .parseSignedClaims(token);
-            return true;
-        } catch (io.jsonwebtoken.security.SecurityException | MalformedJwtException e) {
-            log.info("잘못된 JWT 서명입니다.");
-        } catch (ExpiredJwtException e) {
-            log.info("만료된 JWT 토큰입니다.");
-        } catch (UnsupportedJwtException e) {
-            log.info("지원되지 않는 JWT 토큰입니다.");
-        } catch (IllegalArgumentException e) {
-            log.info("JWT 토큰이 잘못되었습니다.");
-        }
-        return false;
-    }
 }
