+++ src/main/java/com/sascom/chickenstock/global/jwt/JwtResolver.java
@@ -2,12 +2,10 @@ package com.sascom.chickenstock.global.jwt;
 
 import com.sascom.chickenstock.domain.member.entity.Member;
 import com.sascom.chickenstock.global.oauth.dto.MemberPrincipalDetails;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.JwtParser;
-import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.*;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
-import jakarta.annotation.PostConstruct;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -17,6 +15,7 @@ import org.springframework.stereotype.Component;
 import javax.crypto.SecretKey;
 import java.util.Date;
 
+@Slf4j
 @Component
 public class JwtResolver {
     private final JwtProperties jwtProperties;
@@ -66,5 +65,25 @@ public class JwtResolver {
     private Claims getClaims(String token) {
         JwtParser parser = Jwts.parser().verifyWith(secretKey).build();
         return parser.parseSignedClaims(token).getPayload();
+    }
+
+    public boolean isValidToken(String token) {
+
+        try {
+            Jwts.parser()
+                    .verifyWith(secretKey)
+                    .build()
+                    .parseSignedClaims(token);
+            return true;
+        } catch (io.jsonwebtoken.security.SecurityException | MalformedJwtException e) {
+            log.info("잘못된 JWT 서명입니다.");
+        } catch (ExpiredJwtException e) {
+            log.info("만료된 JWT 토큰입니다.");
+        } catch (UnsupportedJwtException e) {
+            log.info("지원되지 않는 JWT 토큰입니다.");
+        } catch (IllegalArgumentException e) {
+            log.info("JWT 토큰이 잘못되었습니다.");
+        }
+        return false;
     }
 }
